Renaming some functions for better understanding and created TRY function for all requests
