+++ app/controllers/categories/all.js
@@ -1,12 +1,10 @@
 const categories = require('../../models/categories/')
-  , { del } = categories()
+  , { all } = categories()
 
 // GETTING CATEGORIES LIST
 module.exports = async (req, res, next) => {
-  const { id } = req.params
-
   req.try(
-    () => del(id)
+    () => all()
   )
 
   next()
