+++ app/controllers/categories/save.js
@@ -1,14 +1,13 @@
 const categories = require('../../models/categories/')
-  , { get } = categories()
+  , { save } = categories()
 
 // GETTING CATEGORIES LIST
 module.exports = async (req, res, next) => {
-  try {
-    res.send(
-      await get()
-    )
-    next()
-  } catch (e) {
-    req.error(e)
-  }
+  const data = req.params
+
+  req.try(
+    () => save(data)
+  )
+
+  next()
 }
