+++ app/controllers/categories/update.js
@@ -1,15 +1,13 @@
 const categories = require('../../models/categories/')
-  , { put } = categories()
+  , { update } = categories()
 
 // GETTING CATEGORIES LIST
 module.exports = (req, res, next) => {
   const { data, id } = req.params
 
-  console.log(req.params)
-
-  put(data, id)
-    .then(data => res.send(data))
-    .catch(e => res.send(e))
+  req.try(
+    () => update(data, id)
+  )
 
   next()
 }
