+++ app/models/categories/all.js
@@ -2,8 +2,8 @@ module.exports = Connection =>
   new Promise((resolve, reject) =>
     Connection
       .then(DB =>
-        DB.query('SELECT * FROM categoriess', (e, data) =>
-          e ? reject('lala') : resolve(data)
+        DB.query('SELECT * FROM categories', (e, data) =>
+          e ? reject(e) : resolve(data)
         )
       )
       .catch(e => reject(e))
